backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
supabase==2.0.2
httpx<0.25.0,>=0.24.0
beautifulsoup4==4.12.2
requests==2.31.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic[email]==2.5.0
passlib==1.7.4
bcrypt==4.0.1
itsdangerous==2.1.2
pandas==2.0.3
lxml==4.9.3